added scaladoc, refactored and insert predisposition to draw food
