--- res/serialization/speciesTest.txt
@@ -1,8 +1,0 @@
-{
-  "icon": "",
-  "name": "TestName",
-  "size": "Small",
-  "strength": 10,
-  "sight": 10,
-  "alimentationType": "Carnivore"
-}
