+++ src/main/scala/view/SimulationGui.scala
@@ -1,11 +1,17 @@
 package view
 
-import model.{Animal, Habitat}
+import model.{Animal, FoodInstance, Habitat}
 
 import javax.swing.WindowConstants
 import scala.swing.{Dimension, Frame, SimpleSwingApplication}
 
-class SimulationGui(habitat: Habitat, val shapePanel: ShapePanel) extends SimpleSwingApplication {
+/**
+ * Class used to create the window where show the simulation.
+ *
+ * @param habitat the Habitat of the simulation.
+ * @param simulationPanel the Panel where to draw all the element of the simulation (areas, animals and food).
+ */
+class SimulationGui(habitat: Habitat, val simulationPanel: SimulationPanel) extends SimpleSwingApplication {
 
   val (width, height) = habitat.dimensions
 
@@ -13,15 +19,22 @@ class SimulationGui(habitat: Habitat, val shapePanel: ShapePanel) extends Simple
     this.resizable = false
     this.size = new Dimension(width, height)
     this.title = "Simulation"
-    contents = shapePanel
+    contents = simulationPanel
     peer.setDefaultCloseOperation(WindowConstants.EXIT_ON_CLOSE)
   }
 
-  def updatePanel(animalsAndRectangles: Map[Animal, Rectangle]): Unit = {
-    shapePanel.peer.removeAll()
-    shapePanel.drawHabitat(habitat)
-    shapePanel.drawAnimals(animalsAndRectangles)
-    shapePanel.revalidate()
-    shapePanel.repaint()
+  /**
+   * Method use to delete the old element in the Panel and draw on the Panel the new state of the simulation.
+   *
+   * @param animalsAndRectangles a Map containing the info of the Animals and Rectangles used to draw them.
+   * @param food the food to draw on the Panel.
+   */
+  def updatePanel(animalsAndRectangles: Map[Animal, Rectangle], food: Seq[FoodInstance]): Unit = {
+    simulationPanel.peer.removeAll()
+    simulationPanel.drawHabitat(habitat)
+    simulationPanel.drawAnimals(animalsAndRectangles)
+    simulationPanel.drawFood(food)
+    simulationPanel.revalidate()
+    simulationPanel.repaint()
   }
 }
